Fix compilation error with analytics on arm64-v8
It seems that CLang on 64 bit arm is a ltitle bit nitpicking about
types. Using good old `static_cast` helps.

+++ src/qtfirebaseservice.cpp
@@ -182,7 +182,7 @@ QVariant QtFirebaseService::fromFirebaseVariant(const firebase::Variant &v)
         case firebase::Variant::kTypeNull:
             return QVariant();
         case firebase::Variant::kTypeInt64:
-            return QVariant(v.int64_value());
+            return QVariant(static_cast<qint64> (v.int64_value()));
         case firebase::Variant::kTypeDouble:
             return QVariant(v.double_value());
         case firebase::Variant::kTypeBool:
